backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.2
pinecone>=3.2.2
openai>=1.30.1
anthropic>=0.28.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
requests>=2.32.3
python-dotenv>=1.0.1
pydantic>=2.7.1
youtube-transcript-api>=0.6.2
PyMuPDF>=1.24.3
aiohttp>=3.9.5
lxml>=5.2.2
psycopg2-binary>=2.9.9
